* src/hd-switcher-menu.c:         - Implemented jumping from pane to pane. Unfortunately you cannot jump         to menu item at same height.         * libhildondesktop/hildon-desktop-popup-menu.[ch]:         - Added select first item method.         - If item to select is NULL, deselect everything.         * libhildondesktop/hildon-desktop-popup-window.c:         - "You shall not compare integers and unsigned integers without a         casting". 	* ChangeLog updated.
git-svn-id: https://stage.maemo.org/svn/maemo/projects/haf/trunk@11582 50eb543b-83fa-0310-9bb7-f7cbbee4ab39

+++ hildon-desktop/libhildondesktop/hildon-desktop-popup-window.c
@@ -901,18 +901,18 @@ void
 hildon_desktop_popup_window_jump_to_pane (HildonDesktopPopupWindow *popup, gint pane)
 {
   g_assert (HILDON_DESKTOP_IS_POPUP_WINDOW (popup));
-
-  if (pane >= popup->priv->n_extra_panes)
+  
+  if (pane >= (gint) popup->priv->n_extra_panes)
     pane = popup->priv->n_extra_panes -1;	  
-	
+  
   if (GTK_WIDGET_VISIBLE (GTK_WIDGET (popup)))
   {
     if (pane <= -1)
-    {
+    { 
       if (popup->priv->have_xgrab)	    
         return;
       else
-      {
+      { 
         gtk_grab_remove (popup->priv->pane_with_grab);	      
 
 	popup_grab_on_window (GTK_WIDGET (popup)->window, GDK_CURRENT_TIME, TRUE);
@@ -924,7 +924,7 @@ hildon_desktop_popup_window_jump_to_pane (HildonDesktopPopupWindow *popup, gint 
       }
     }
     else
-    {
+    { 
       GtkWidget *widget_with_grab;
 	    
       if (popup->priv->have_xgrab)
